requirements.txt
# Usage: pip install -r requirements.txt
Flask
Flask-RESTful
python-decouple
flask-cors
python-dotenv
# Base ----------------------------------------
numpy
onnxruntime
matplotlib
opencv-python